--- back/src/modules/snowball/dto/res-decoration-snowball.dto.ts
@@ -1,19 +1,0 @@
-import { ValidateNested, IsUUID } from '@nestjs/class-validator';
-import { ApiProperty } from '@nestjs/swagger';
-
-export class DecorationSnowballDto {
-  @IsUUID()
-  @ApiProperty({
-    type: String,
-    description: '업데이트된 스노우볼 uuid',
-    format: 'uuid'
-  })
-  readonly snowball_id: string;
-
-  @ValidateNested({ each: true })
-  @ApiProperty({
-    type: [DecorationSnowballDto],
-    description: '업데이트된 데코오브젝트들이 들어있는 리스트'
-  })
-  readonly deco_list: DecorationSnowballDto[];
-}
